pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-agent-boilerplate"
version = "0.1.0"
description = "AI Agent Boilerplate using FastAPI + PydanticAI + OpenAI + Langfuse + Celery"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "agents_core"}, {include = "apps"}, {include = "workers"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic-ai = "^0.0.14"
openai = "^1.40.0"
langfuse = "^2.36.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
celery = "^5.3.4"
redis = "^5.0.1"
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agents_core", "apps", "workers"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
